+++ lib/core/router/router_manager.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/cupertino.dart';
 import 'package:go_router/go_router.dart';
 import 'package:words_learning/core/router/routes.dart';
-import 'package:words_learning/features/courses/screens/all_courses_screen.dart';
-import 'package:words_learning/features/courses/screens/course_screen.dart';
-import 'package:words_learning/features/courses/screens/my_courses_screen.dart';
-import 'package:words_learning/features/learning/screens/deep_learning_screen.dart';
-import 'package:words_learning/features/learning/screens/drag_and_drop_screen.dart';
-import 'package:words_learning/features/learning/screens/test_screen.dart';
-import 'package:words_learning/features/learning/screens/matching_screen.dart';
-import 'package:words_learning/features/learning/screens/writing_screen.dart';
+import 'package:words_learning/features/courses/presentation/screens/all_courses_screen.dart';
+import 'package:words_learning/features/courses/presentation/screens/course_screen.dart';
+import 'package:words_learning/features/courses/presentation/screens/my_courses_screen.dart';
+import 'package:words_learning/features/learning/presentation/screens/deep_learning_screen.dart';
+import 'package:words_learning/features/learning/presentation/screens/drag_and_drop_screen.dart';
+import 'package:words_learning/features/learning/presentation/screens/test_screen.dart';
+import 'package:words_learning/features/learning/presentation/screens/matching_screen.dart';
+import 'package:words_learning/features/learning/presentation/screens/writing_screen.dart';
 import 'package:words_learning/features/profile/screens/profile_screen.dart';
 
 final class RouterManager {
