+++ lib/features/courses/presentation/screens/my_courses_screen.dart
@@ -2,19 +2,28 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:words_learning/common/widgets/light_grey_divider.dart';
 import 'package:words_learning/core/router/routes.dart';
-import 'package:words_learning/features/courses/widgets/main_course_widget.dart';
-import 'package:words_learning/features/courses/widgets/secondary_course_widget.dart';
+import 'package:words_learning/core/theme/color_palette.dart';
+import 'package:words_learning/features/courses/presentation/widgets/main_course_widget.dart';
+import 'package:words_learning/features/courses/presentation/widgets/secondary_course_widget.dart';
+import 'package:words_learning/features/courses/presentation/widgets/streak_widget.dart';
 
-class AllCoursesScreen extends StatelessWidget {
-  const AllCoursesScreen({super.key});
+class MyCoursesScreen extends StatelessWidget {
+  const MyCoursesScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
     final TextTheme textTheme = Theme.of(context).textTheme;
     return Scaffold(
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          // todo: add navigation to AddCourseScreen()
+        },
+        backgroundColor: ColorPalette.mainFocusColor,
+        child: Icon(Icons.add),
+      ),
       appBar: AppBar(
         title: Text(
-          'ALL COURSES',
+          'MY COURSES',
           style: textTheme.titleLarge,
         ),
         actions: [
@@ -39,13 +48,16 @@ class AllCoursesScreen extends StatelessWidget {
               LightGreyDivider(),
               SizedBox(height: 16),
               MainCourseWidget(
-                  onTap: () {
-                    context.go(Routes.course.path);
-                  },
-                  textTheme: textTheme,
-                  courseTitle: 'Swear words',
-                  courseDescription:
-                      'Master the art of offensive and profane language.'),
+                textTheme: textTheme,
+                courseTitle: 'Swear words',
+                courseDescription:
+                    'Master the art of offensive and profane language.',
+                onTap: () {
+                  context.go(Routes.course.path);
+                },
+              ),
+              SizedBox(height: 16),
+              StreakWidget(textTheme: textTheme),
               SizedBox(height: 16),
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -63,15 +75,6 @@ class AllCoursesScreen extends StatelessWidget {
                       textTheme: textTheme),
                 ],
               ),
-              SizedBox(height: 12),
-              MainCourseWidget(
-                  onTap: () {
-                    context.go(Routes.course.path);
-                  },
-                  textTheme: textTheme,
-                  courseTitle: 'Swear words',
-                  courseDescription:
-                      'Master the art of offensive and profane language.'),
               SizedBox(height: 12),
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
