+++ lib/features/courses/presentation/widgets/main_course_widget.dart
