+++ lib/features/courses/presentation/widgets/streak_widget.dart
