+++ lib/features/learning/presentation/screens/drag_and_drop_screen.dart
@@ -202,7 +202,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:words_learning/common/widgets/rectangle_icon_button.dart';
 import 'package:words_learning/core/theme/color_palette.dart';
-import 'package:words_learning/features/learning/widgets/word_button.dart';
+import 'package:words_learning/features/learning/presentation/widgets/word_button.dart';
 
 class DragAndDropScreen extends StatefulWidget {
   const DragAndDropScreen({super.key});
